Pare down CI targets
geerlingguy.docker seems to have relatively similar CI, but tests
against fewer targets and seems to be less complex over all, and has
successful action runs. this is an attempt to pare our stuff down and
see if that helps. we probably should not really be relying on older
python and ansible versions to begin with.