CI: improve with uv (#17)
* ci(actions): introduce uv to linting workflow

* ci(actions): introduce uv to main testing workflow

Uses a different approach to caching (at least for now) based on (https://docs.astral.sh/uv/guides/integration/github/#caching).

* ci(actions): move cache restore to after requirements.txt

* ci(actions): use env var for cache dir instead of restating value

* ci(actions): try only installing requirements if no cache hit

See https://github.com/actions/cache?tab=readme-ov-file#skipping-steps-based-on-cache-hit

* Revert "ci(actions): try only installing requirements if no cache hit"

This reverts commit 03d57d4ef5d549b471091a96e9ed1db532dca54c.

* ci(actions): try moving requirements generation before uv install

...to see if that resolves most of the remaining warnings.

* ci(actions): update uv to 0.6.4

* ci(actions): try removing extraneous problem matcher

In the lint job, the uv setup task complains about there not being a
requirements.txt or the like. We don't have one in the linting step
since this isn't a formal Python project, so this warning can likely be
safely suppressed.

* Revert "ci(actions): try removing extraneous problem matcher"

This reverts commit 00ef682de82bc61abdd31e6162ffa7976742369c.

The previous commit seemed to stop all of the actions that we wanted to
run from running, which isn't good!